pyproject.toml
# ── pyproject.toml ──
[build-system]
requires      = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "texttoknowledgegraph"
version         = "0.3.3"
description     = "A Python package to generate BEL statements and CX2 networks."
readme          = "textToKnowledgeGraph/README.md"
requires-python = ">=3.11"

authors = [
  { name = "Favour James", email = "favour.ujames196@gmail.com" },
]

dependencies = [
  "langchain==0.3.25",
  "langchain_core==0.3.60",
  "langchain_openai==0.3.17",
  "lxml==5.2.1",
  "ndex2>=3.8.0,<4.0.0",
  "pandas",
  "openai==1.79.0",
  "pydantic==2.11.4",
  "pytest==7.4.4",
  "python-dotenv==1.1.0",
  "Requests==2.32.3",
  "gilda==1.4.0",
  "markitdown==0.1.1"
]

classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

[project.scripts]
textToKnowledgeGraph = "textToKnowledgeGraph.main:cli"

[project.urls]
Homepage = "https://github.com/ndexbio/llm-text-to-knowledge-graph"

# ────────── setuptools-specific knobs ──────────
[tool.setuptools]
packages             = { find = {} }
include-package-data = true       

[tool.setuptools.package-data]
"textToKnowledgeGraph" = ["*.txt", "*.json"]
